% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df-functions.R
\name{add_heading}
\alias{add_heading}
\title{Add headings}
\usage{
add_heading(X)
}
\arguments{
\item{X}{A data frame with columns \code{ymd}, \code{time_label} and \code{tex}.}
}
\value{
A data frame with an extra column \code{tex_w_heading}.
}
\description{
Given a data frame with columns \code{ymd}, \code{time_label} and \code{tex} return a data
frame with
the same columns but with an addition \code{tex_w_heading} column made from the \code{ymd}
and \code{time_label} columns.
}
