% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment-time-label.R
\name{augment_time_label}
\alias{augment_time_label}
\title{Add Time Labels to a Data Frame of Text}
\usage{
augment_time_label(X)
}
\arguments{
\item{X}{A data frame with a column named text.}
}
\value{
A data frame with new columns, ts_s, ts, time_label.
}
\description{
Add Time Labels to a Data Frame of Text
}
