% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-time-label.R
\name{extract_time_label}
\alias{extract_time_label}
\title{Match, Parse and Format Timestamps from Text}
\usage{
extract_time_label(text)
}
\arguments{
\item{text}{Input text.}
}
\value{
A formatted timestamp string.
}
\description{
This function reads and parses timestamps from text and then outputs a
differently formatted string representing the same timestamp.
}
