% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gqg.R
\name{gqg}
\alias{gqg}
\title{Chop a String of Text to a Specific Width}
\usage{
gqg(x, textwidth)
}
\arguments{
\item{x}{A string.}

\item{textwidth}{Desired width of output strings.}
}
\value{
A string with more newline characters.
}
\description{
Chop a String of Text to a Specific Width
}
