% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gqg-paragraph.R
\name{gqg_paragraph}
\alias{gqg_paragraph}
\title{Chop a String of Text to a Specific Width}
\usage{
gqg_paragraph(s, textwidth)
}
\arguments{
\item{textwidth}{}
}
\description{
Chop a String of Text to a Specific Width
}
