% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-heading.R
\name{latex_section_heading}
\alias{latex_section_heading}
\title{Create a LaTeX Section Heading}
\usage{
latex_section_heading(ymd, time_s)
}
\arguments{
\item{ymd}{A date in ymd format.}

\item{time_s}{A time string.}
}
\value{
A LaTeX section heading string.
}
\description{
Given a date in ymd format (e.g. 2024/09/01) and time string (e.g. 06:06 am)
this function returns a LaTeX section heading.
}
\examples{
latex_section_heading("2024/09/01", "06:06 am")
}
