% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repair-latex.R
\name{repair_latex}
\alias{repair_latex}
\title{Repair LaTeX}
\usage{
repair_latex(X)
}
\arguments{
\item{X}{A data frame with a text column.}
}
\value{
A new data frame containing a tex column.
}
\description{
Repair LaTeX
}
