% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repair-latex-string.R
\name{repair_latex_string}
\alias{repair_latex_string}
\title{Escape A Few Things}
\usage{
repair_latex_string(x)
}
\arguments{
\item{x}{A text strirng.}
}
\value{
A text string with some escaped characters.
}
\description{
Replace certain characters in a string with escaped versions of those same
characters for the benefit of LaTeX.
}
