% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace-at-i.R
\name{replace_at_i}
\alias{replace_at_i}
\title{Replace a Character at a Location in a String}
\usage{
replace_at_i(s, i, x = "\\n")
}
\arguments{
\item{x}{}
}
\description{
Replace a Character at a Location in a String
}
