% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strip-time-headings.R
\name{strip_time_headings}
\alias{strip_time_headings}
\title{Remove Datetime Strings from Text}
\usage{
strip_time_headings(text)
}
\arguments{
\item{text}{A string of text, possibly containing datetime headings.}
}
\value{
The input string with all datetime headings removed.
}
\description{
Removes all datetime headings like "\% Tue 3 Mar 09:20:25 GMT 2020".
}
