% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ymd-text-format.R
\name{ymd_text_format}
\alias{ymd_text_format}
\title{Convert a Date Into Text.}
\usage{
ymd_text_format(ymd)
}
\arguments{
\item{ymd}{A date in ymd format.}
}
\value{
A date string.
}
\description{
This takes a ymd (e.g. 2024/09/01) and returns the same date in text format
(e.g. Sunday 1 September 2024).
}
