% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate-last-space-before.R
\name{locate_last_space_before}
\alias{locate_last_space_before}
\title{Locate Last Space Before a Character in a String}
\usage{
locate_last_space_before(s, x)
}
\arguments{
\item{x}{}
}
\description{
Locate Last Space Before a Character in a String
}
