% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace-at-i.R
\name{replace_at_i}
\alias{replace_at_i}
\title{Replace a Character at a Location in a String}
\usage{
replace_at_i(s, i, x = "\\n")
}
\arguments{
\item{s}{A string.}

\item{i}{An index.}

\item{x}{A character.}
}
\value{
The string s but with character x inserted at position i.
}
\description{
Replace a Character at a Location in a String
}
